caricature-visual-face-match-project/python_service/requirements.txt
# Python Backend Service Dependencies
# Core
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0

# Face Detection (YOLOv5-Face)
scikit-image>=0.18.0
PyYAML>=5.3.1
scipy>=1.4.1
tqdm>=4.41.0
thop>=0.1.1
requests>=2.25.0
matplotlib>=3.3.0
pandas>=1.1.0

# Cross-modal matching
git+https://github.com/openai/CLIP.git
facenet-pytorch>=2.5.0

# API Server
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Utilities
pydantic>=1.8.0
python-dotenv>=0.19.0
